requirements.txt
python-dotenv
Flask
openai
echo "requests" 